+++ Utils/RandomValueGenerator.hpp
@@ -7,14 +7,14 @@ namespace DIM {
     class RandomValueGenerator {
       private:
         RandomValueGenerator();
-        static RandomValueGenerator *instance;
+        static RandomValueGenerator* instance;
 
       public:
         ~RandomValueGenerator();
         static const RandomValueGenerator* getInstance();
         bool getRandomBool(unsigned short probability) const;
         int getRandomIntInRange(int lowest, int highest) const;
-        float getRandomFloatInRange(float min, int max) const;
+        float getRandomFloatInRange(float min, float max) const;
     };
     
   }
